refactor: parking management service with enhanced webhook handling and database operations
- Introduced `PersistenceHandler` for centralized database operation handling with error logging.
- Updated `ParkingSpotRepositoryAdapter` to utilize `PersistenceHandler` for finding parking spots.
- Changed `ParkingEntity` and `ParkingEventEntity` to use `BigDecimal` for monetary values.
- Created `ParkingCustomQueryRepository` and its implementation for custom database queries related to parking.
- Refactored `ParkingEventRepository` to include new query methods for event retrieval.
- Implemented new webhook handlers (`EntryWebhookHandler`, `ParkedWebhookHandler`, `ExitWebhookHandler`) for processing parking events.
- Consolidated webhook handling logic into `ParkingWebhookHandler`.
- Added `CalculatePricingMultiplierHandler` for dynamic pricing calculations based on occupancy.
- Removed the old `WebhookService` and replaced it with a more modular approach using specific handlers.
- Updated database schema to accommodate new fields for pricing in `parking_events`.
- Enhanced exception handling with new messages for better user feedback.

+++ src/main/kotlin/io/github/martinsjavacode/parkingmanagement/infra/persistence/entity/ParkingEntity.kt
@@ -3,6 +3,7 @@ package io.github.martinsjavacode.parkingmanagement.infra.persistence.entity
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
 import org.springframework.data.relational.core.mapping.Table
+import java.math.BigDecimal
 import java.time.LocalTime
 
 @Table("parking")
@@ -12,7 +13,7 @@ class ParkingEntity(
     @Column("sector_name")
     var sector: String,
     @Column("base_price")
-    var basePrice: Double,
+    var basePrice: BigDecimal,
     @Column("max_capacity")
     var maxCapacity: Int,
     @Column("open_hour")
